added layout, moved demo to palette
